Pluralize the PackageSymlinkInfo provider
This change modifies the PackageSymlinkInfo to make it contain a dictionary of
destination -> src mappings instead simply being one-to-one.

This is to allow rules such as `pkg_mklinks` to emit collections of
symlinks, rather than requiring users to make one rule for each and every
symlink they may have in their package.

Additionally, the documentation of `PackageFilegroupInfo` was also updated to
properly name `PackageDirsInfo`.